refactor auth module

+++ src/modules/users/application/use-cases/ban-unban-user.handler.ts
@@ -1,5 +1,7 @@
+import { BadRequestException } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
+import { IUserBanRepository } from '../../../bans/infrastructure/user-bans.repository.';
 import { ISecurityRepository } from '../../../security/infrastructure/security.sql.repository';
 import { IUsersRepository } from '../../infrastructure/users.repository.sql';
 import { BanUnbanUserInput } from '../dto/ban-unban-user.input';
@@ -16,12 +18,21 @@ export class BanUnbanUserHandler
   implements ICommandHandler<BanUnbanUserCommand>
 {
   constructor(
-    private readonly usersRepository: IUsersRepository,
+    private readonly UserBanRepository: IUserBanRepository,
     private readonly securityRepository: ISecurityRepository,
   ) {}
 
-  async execute({ userId, payload }: BanUnbanUserCommand) {
+  async execute({ userId, payload }: BanUnbanUserCommand): Promise<boolean> {
+    const isUserBanned = await this.UserBanRepository.createOrUpdateBan(
+      userId,
+      payload,
+    );
+    if (!isUserBanned) {
+      throw new BadRequestException();
+    }
+
     await this.securityRepository.removeAllDevices(userId);
-    return this.usersRepository.createOrUpdateBan(userId, payload);
+
+    return true;
   }
 }
